Changes to integration step size, parallel demo
Set the maximum step size that the integrator can take to the dt
specified by the model input. This is mostly necessary for functions
like ‘pulse’ which are discontinuous, as the integrator needs to know
to look in their integration window.

Added a demo of using ipython parallelization to speed up Monte Carlo
analysis.