Addressed reset_state issues
What had happened was that we moved the state vector setup (that
creates an empty dictionary) into the `reset_state` function. But, we
had called `reset_state` recursively to handle the situations where one
variable depended on another. To resolve the issue, created a
sub-function specifically for the recursive part.
Seems to resolve the issue, and passes the [chained initialization
test](https://github.com/SDXorg/test-models/tree/master/tests/chained_in
itialization) written for this type of situation.

Still fails a bunch of other tests…