Resolving conflicts
Rebased onto master, munged some bits that didn’t rebase properly,
worked skipped some tests that are left over from my sloppy merging of
tests branch.

There is an issue with `delay N` in that the results are slightly
different from cannon, and I’m not sure why this should be the case.