Initial function fix
If the function in Vensim is defined as:
initial var=
	INITIAL(var)
Then the builder creates a new variable initial_var, which is the same as the already existing variable.
In the python file this is then solved by:
Return utils.xrmerge([
	Initial_var(),
	Functions.Initial(lambda: var()),
])
When executing the python file, this creates a "maximum recursion depth exceeded" error.

Changin the placeholder to "initializer" does not solve the problem, but makes it less likely to occur ( in all vensim models on the society homepage, the word "initializer" is never used)