Reorder xarrays dims to mantain original dim
Reorder xarrays dims to mantain original dim. Some functions, for example when constructing the inputs for stateful elements, change the order of the dimensions in the xarray. Reordering the xarrays lets using the coordinate value directly to extract the data from the xarray. This makes possible to run models with the same coordinate values for different dimensions without problem, such as subscript_switching.